Rename parent_title_id to series_title_id in demo.silver.title_episode

+++ transform/silver/models/silver/person.sql
@@ -3,10 +3,10 @@ SELECT
   primaryName AS name,
   CASE 
     WHEN birthYear = '\\N' THEN NULL
-    ELSE CAST(birthYear AS SMALLINT)
+    ELSE CAST(birthYear AS INT)
   END AS birth_year,
   CASE 
     WHEN deathYear = '\\N' THEN NULL
-    ELSE CAST(deathYear AS SMALLINT)
+    ELSE CAST(deathYear AS INT)
   END AS death_year
 FROM {{ source('bronze', 'name_basics') }}
